chore: rename devopsguru construct file
